requirements.txt
pandas
numpy
matplotlib
seabron
flask
sentence-transformers
fastapi
uvicorn
faiss-cpu
